root_path added to build config and file locations
root_path is added as a property of BuildConfig and this is fed into all uses of FileLocation.

Also added a new flag to the CLI. If present, the root_path is set as the value following the flag, if not, the DFM_ROOT_PATH env var is used. If neither are present, the root path defaults to the root of the PC's drive (e.g c:// or / based on OS)